Rename `pulse` -> `capsule`.
