+++ android/app/src/main/kotlin/io/github/reactivecircus/kstreamlined/android/KSActivity.kt
@@ -34,7 +34,7 @@ import androidx.navigation3.runtime.rememberNavBackStack
 import androidx.navigation3.runtime.rememberSaveableStateHolderNavEntryDecorator
 import androidx.navigation3.ui.NavDisplay
 import io.github.reactivecircus.kstreamlined.android.core.designsystem.foundation.KSTheme
-import io.github.reactivecircus.kstreamlined.kmp.pulse.metro.LocalPresenterFactory
+import io.github.reactivecircus.kstreamlined.kmp.capsule.inject.LocalPresenterFactory
 import io.github.reactivecircus.kstreamlined.kmp.settings.model.AppSettings
 
 class KSActivity : ComponentActivity() {
