+++ android/app/src/main/kotlin/io/github/reactivecircus/kstreamlined/android/di/AppGraph.kt
@@ -19,11 +19,11 @@ import io.github.reactivecircus.kstreamlined.android.core.scheduledwork.KSWorker
 import io.github.reactivecircus.kstreamlined.android.core.scheduledwork.sync.SyncScheduler
 import io.github.reactivecircus.kstreamlined.android.licentia.AllLicensesInfo
 import io.github.reactivecircus.kstreamlined.kmp.appinfo.AppInfo
+import io.github.reactivecircus.kstreamlined.kmp.capsule.inject.PresenterGraph
 import io.github.reactivecircus.kstreamlined.kmp.database.FeedItemEntity
 import io.github.reactivecircus.kstreamlined.kmp.database.InstantAdapter
 import io.github.reactivecircus.kstreamlined.kmp.database.KStreamlinedDatabase
 import io.github.reactivecircus.kstreamlined.kmp.database.LastSyncMetadata
-import io.github.reactivecircus.kstreamlined.kmp.pulse.metro.PresenterGraph
 import io.github.reactivecircus.kstreamlined.kmp.settings.datasource.SettingsDataSource
 import io.github.reactivecircus.kstreamlined.kmp.settings.datasource.datastore.createAppSettingsDataStore
 import io.github.reactivecircus.licentia.runtime.LicensesInfo
