+++ kmp/capsule/inject/build.gradle.kts
@@ -1,0 +1,19 @@
+import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
+
+kstreamlined {
+    kmpLibrary {
+        targets {
+            jvm()
+            ios()
+        }
+        compose()
+
+        @OptIn(ExperimentalKotlinGradlePluginApi::class)
+        dependencies {
+            api(project(":kmp:capsule:runtime"))
+            api(libs.metro.runtime)
+            api(libs.androidx.compose.runtime)
+            implementation(libs.androidx.compose.runtime.retain)
+        }
+    }
+}
