+++ kmp/capsule/runtime/src/commonMain/kotlin/io/github/reactivecircus/kstreamlined/kmp/capsule/runtime/PresenterAssistedFactory.kt
@@ -1,0 +1,6 @@
+package io.github.reactivecircus.kstreamlined.kmp.capsule.runtime
+
+/**
+ * Marker interface for assisted factories for [Presenter] types.
+ */
+public interface PresenterAssistedFactory
