--- kmp/pulse/metro/src/commonMain/kotlin/io/github/reactivecircus/kstreamlined/kmp/pulse/metro/PresenterAssistedFactory.kt
@@ -1,6 +1,0 @@
-package io.github.reactivecircus.kstreamlined.kmp.pulse.metro
-
-/**
- * Marker interface for assisted factories for [Presenter] types.
- */
-public interface PresenterAssistedFactory
