--- kmp/pulse/metro/src/commonMain/kotlin/io/github/reactivecircus/kstreamlined/kmp/pulse/metro/PresenterAssistedFactoryKey.kt
@@ -1,21 +1,0 @@
-package io.github.reactivecircus.kstreamlined.kmp.pulse.metro
-
-import dev.zacsweers.metro.MapKey
-import kotlin.reflect.KClass
-
-/**
- * A [MapKey] annotation for binding [PresenterAssistedFactory] instances in a multibinding map.
- */
-@MapKey(implicitClassKey = true)
-@Target(
-    AnnotationTarget.FUNCTION,
-    AnnotationTarget.FIELD,
-    AnnotationTarget.PROPERTY,
-    AnnotationTarget.PROPERTY_GETTER,
-    AnnotationTarget.CLASS,
-    AnnotationTarget.TYPE,
-)
-@Retention(AnnotationRetention.RUNTIME)
-public annotation class PresenterAssistedFactoryKey(
-    val value: KClass<out PresenterAssistedFactory> = Nothing::class,
-)
