+++ settings.gradle.kts
@@ -70,21 +70,16 @@ rootProject.name = "kstreamlined-mobile"
 
 includeKmpProjects(
     ":kmp:app-info",
+    ":kmp:capsule:inject",
+    ":kmp:capsule:runtime",
+    ":kmp:database",
+    ":kmp:database-testing",
+    ":kmp:datastore-testing",
     ":kmp:feed-datasource",
     ":kmp:feed-model",
     ":kmp:feed-sync:common",
     ":kmp:feed-sync:runtime",
     ":kmp:feed-sync:testing",
-    ":kmp:settings-datasource",
-    ":kmp:settings-model",
-    ":kmp:remote:common",
-    ":kmp:remote:cloud",
-    ":kmp:remote:edge",
-    ":kmp:remote:mock",
-    ":kmp:remote:testing",
-    ":kmp:database",
-    ":kmp:database-testing",
-    ":kmp:datastore-testing",
     ":kmp:network-monitor:common",
     ":kmp:network-monitor:runtime",
     ":kmp:network-monitor:testing",
@@ -96,8 +91,13 @@ includeKmpProjects(
     ":kmp:presentation:settings",
     ":kmp:presentation:talking-kotlin-episode",
     ":kmp:pretty-time",
-    ":kmp:pulse:metro",
-    ":kmp:pulse:runtime",
+    ":kmp:remote:common",
+    ":kmp:remote:cloud",
+    ":kmp:remote:edge",
+    ":kmp:remote:mock",
+    ":kmp:remote:testing",
+    ":kmp:settings-datasource",
+    ":kmp:settings-model",
 )
 
 includeAndroidProjects(
